==> 01_Realms_setup.Rmd <==
---
title: "Realms-Setup"
output: html_document
date: "2023-10-06"
editor_options: 
  
  chunk_output_type: console
---
```{r}
#Load data 
rm(list=ls())

#Loading required libraries 

library(tidyverse)
library(devtools)
library(plyr) 
library(Rcpp)
library(reshape2) 
library(inline)
library(ggplot2)
library(ggthemes)
library(mizer)
library(ggpubr)
library(mizerMR)
library(mizerExperimental)
library(readxl)
library(mizerHowTo)
library(mgcv)
library(scales)
library(patchwork)

summarise <-dplyr::summarise

my_theme<-theme_bw()+
  theme(text = element_text(size = 10), # this should be overwritten by the below
        plot.title = element_text(size = 10),
        axis.title = element_text(size = 9),
        axis.text = element_text(size = 6),
        legend.title=element_text(size = 9),
        legend.text=element_text(size = 8),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        legend.key.size = unit(0.1, "cm"))

```


##Workflow follows Phoebe's workflow  

```{r}

Sea_params <- readRDS("final_params_object.rds")
Sea_params@species_params$temp_min <- c(13,14,12,2,2,14,4,11,12,9,11,3,8,9,8,3,14,14,6)
Sea_params@species_params$temp_max <-c(29,20,23,15,16,21,19,16,19,29,20,9,16,16,14,13,18,20,24)

interaction <- Sea_params@interaction
interaction <- as.matrix(interaction)
params <- newMultispeciesParams(species_params = Sea_params@species_params,interaction = interaction, kappa = 0.002191539)

realm_names <- c("r0-300", "r0-500","r0-400","r100-700","r200-1100", "r150-800", "r400-1000", "r400-800", "r400-600", "r100-800", "r100-400", "r50-300", "r150-400", "r50-500", "r50-400" )
species_names <- as.character(Sea_params@species_params$species)
sizes <- Sea_params@w 

# Create the vertical migration array and fill it
vertical_migration_array <- array(0, dim = (c(length(realm_names), length(species_names), length(sizes))), dimnames = list(realm = realm_names, sp = species_names, w = signif(sizes,3))) # realm x species x size

# Indices for realms 
r1 <- which(realm_names == "r0-300") 
r2 <- which(realm_names == "r0-500") 
r3 <- which(realm_names == "r0-400") 
r4 <- which(realm_names == "r100-700")
r5 <- which(realm_names == "r200-1100")
r6<- which(realm_names == "r150-800")
r7<- which(realm_names == "r400-1000")
r8<- which(realm_names == "r400-800")
r9<- which(realm_names == "r400-600")
r10<- which(realm_names == "r100-800")
r11<- which(realm_names == "r100-400")
r12<- which(realm_names == "r50-300")
r13<- which(realm_names == "r150-400")
r14<- which(realm_names == "r50-500")
r15<- which(realm_names == "r50-400")


# Indices for species 
sp1 <- which(species_names == "lanternfish") # 0-300
sp2 <- which(species_names ==  "school whiting")#0-500
sp3 <- which(species_names == "arrow squid")#0-300
sp4 <- which(species_names == "bigeye ocean perch")#150-400
sp5 <- which(species_names == "mackerel")#50-300
sp6 <- which(species_names == "redfish")#100-700
sp7 <- which(species_names == "deepwater sharks")# 200-1100
sp8 <- which(species_names == "jackass morwong")# 150-400
sp9 <- which(species_names == "tiger flathead")# 50-400
sp10 <- which(species_names == "dories and oreos")# 150-800 
sp11 <- which(species_names == "blue warehou" )#50-300
sp12 <- which(species_names == "orange roughy")# 400-1000 
sp13 <- which(species_names == "blue grenadier")# 400-800
sp14 <- which(species_names == "silver warehou")#400-600
sp15 <- which(species_names ==  "gemfish")#100-800
sp16 <- which(species_names == "pink ling")#100-400
sp17 <- which(species_names ==  "common sawshark")#50-500
sp18 <- which(species_names == "gummy shark")#50-400
sp19 <- which(species_names ==  "school shark")#50-500


# Vertical migration array
vertical_migration_array[r1, sp1, ] <- 1
vertical_migration_array[r2, sp2, ] <- 1
vertical_migration_array[r1, sp3, ] <- 1 
vertical_migration_array[r13, sp4, ] <- 1
vertical_migration_array[r12, sp5, ] <- 1 
vertical_migration_array[r4, sp6, ] <- 1
vertical_migration_array[r5, sp7, ] <- 1
vertical_migration_array[r13, sp8, ] <- 1
vertical_migration_array[r15, sp9, ] <- 1
vertical_migration_array[r6, sp10, ] <- 1
vertical_migration_array[r12, sp11, ] <- 1
vertical_migration_array[r7, sp12, ] <- 1
vertical_migration_array[r8, sp13, ] <- 1
vertical_migration_array[r9, sp14, ] <- 1
vertical_migration_array[r10, sp15, ] <- 1
vertical_migration_array[r11, sp16, ] <- 1
vertical_migration_array[r14, sp17, ] <- 1
vertical_migration_array[r15, sp18, ] <- 1
vertical_migration_array[r14, sp19, ] <- 1


# Construct exposure array
exposure_array <- array(0, dim = (c(length(realm_names), length(species_names))), 
                  dimnames = list(realm = realm_names, sp = species_names)) # realm x species

for (r in seq(1,length(realm_names),1)) {
    for (s in seq(1,length(species_names),1)) {
        if (any(vertical_migration_array[r,s,] > 0)) {
            exposure_array[r,s] = 1
        }
    }
}

# Save data for easy future use
save(vertical_migration_array, file = "SEAus_vertical_migration_array.Rdata")
write.table(exposure_array, file = "SEAus_exposure_array.dat", quote = FALSE, row.names = TRUE, col.names = TRUE)

hgg

# area paramter in Cami's model is based on polygon on google map and it was a much wider polygon - multiplied the area by max depth (area parameter is larger ) - change the area according to the input.

# go through calibration again with the new area ?

```
