==> 03_Prep_Temp_SEA.Rmd <==
---
title: "03_Prep_Temp_SEA"
output: html_document
date: "2024-03-12"
editor_options: 
  chunk_output_type: console
---

```{r}

#Phoebe's Temp workflow 
## Prep temperature forcing for input into therMizer

#This script uses the files output by RealmTemp_ts_gfdl.jnl and RealmTemp_ts_WOA.jnl (for use with [PyFerret](https://ferret.pmel.noaa.gov/Ferret/)) 
#to create an `ocean_temp` array that will be used by therMizer.  These are temperatures that have been vertically averaged over each species#'depth range and then spatially averaged over the model domain.  For details on this, see [Woodworth-Jefcoats et al. 2019](https://www.frontiersin.org/articles/10.3389/fmars.2019.00383/full).

```{r}
# Load WOA baseline, skipping the 6-line header
#r1_WOA <- read.delim("WOA_TEMP_0to20_ave.dat", skip = 6, header = FALSE)
#r2_WOA <- read.delim("WOA_TEMP_0to50_ave.dat", skip = 6, header = FALSE)
#r3_WOA <- read.delim("WOA_TEMP_0to85_ave.dat", skip = 6, header = FALSE)
#r4_WOA <- read.delim("WOA_TEMP_0to100_ave.dat", skip = 6, header = FALSE)
#r5_WOA <- read.delim("WOA_TEMP_0to150_ave.dat", skip = 6, header = FALSE)
#r6_WOA <- read.delim("WOA_TEMP_0to200_ave.dat", skip = 6, header = FALSE)
#r7_WOA <- read.delim("WOA_TEMP_0to1200_ave.dat", skip = 6, header = FALSE)
#r8_WOA <- read.delim("WOA_TEMP_50to150_ave.dat", skip = 6, header = FALSE)
#r9_WOA <- read.delim("WOA_TEMP_50to200_ave.dat", skip = 6, header = FALSE)
#r10_WOA <- read.delim("WOA_TEMP_50to300_ave.dat", skip = 6, header = FALSE)
#r11_WOA <- read.delim("WOA_TEMP_100to200_ave.dat", skip = 6, header = FALSE)
#r12_WOA <- read.delim("WOA_TEMP_100to250_ave.dat", skip = 6, header = FALSE)
#r13_WOA <- read.delim("WOA_TEMP_100to400_ave.dat", skip = 6, header = FALSE)
#r14_WOA <- read.delim("WOA_TEMP_150to600_ave.dat", skip = 6, header = FALSE)
#r15_WOA <- read.delim("WOA_TEMP_200to400_ave.dat", skip = 6, header = FALSE)
#r16_WOA <- read.delim("WOA_TEMP_200to980_ave.dat", skip = 6, header = FALSE)
#r17_WOA <- read.delim("WOA_TEMP_250to500_ave.dat", skip = 6, header = FALSE)
#r18_WOA <- read.delim("WOA_TEMP_400to1200_ave.dat", skip = 6, header = FALSE)

# Load control climate for differencing baseline, skipping the 7-line header
#r1_CTRL <- read.delim("CTRL_THETA_0to20_ave.dat", skip = 7, header = FALSE)
#r2_CTRL <- read.delim("CTRL_THETA_0to50_ave.dat", skip = 7, header = FALSE)
#r3_CTRL <- read.delim("CTRL_THETA_0to85_ave.dat", skip = 7, header = FALSE)
#r4_CTRL <- read.delim("CTRL_THETA_0to100_ave.dat", skip = 7, header = FALSE)
#r5_CTRL <- read.delim("CTRL_THETA_0to150_ave.dat", skip = 7, header = FALSE)
#r6_CTRL <- read.delim("CTRL_THETA_0to200_ave.dat", skip = 7, header = FALSE)
#r7_CTRL <- read.delim("CTRL_THETA_0to1200_ave.dat", skip = 7, header = FALSE)
#r8_CTRL <- read.delim("CTRL_THETA_50to150_ave.dat", skip = 7, header = FALSE)
#r9_CTRL <- read.delim("CTRL_THETA_50to200_ave.dat", skip = 7, header = FALSE)
#r10_CTRL <- read.delim("CTRL_THETA_50to300_ave.dat", skip = 7, header = FALSE)
#r11_CTRL <- read.delim("CTRL_THETA_100to200_ave.dat", skip = 7, header = FALSE)
#r12_CTRL <- read.delim("CTRL_THETA_100to250_ave.dat", skip = 7, header = FALSE)
#r13_CTRL <- read.delim("CTRL_THETA_100to400_ave.dat", skip = 7, header = FALSE)
#r14_CTRL <- read.delim("CTRL_THETA_150to600_ave.dat", skip = 7, header = FALSE)
#r15_CTRL <- read.delim("CTRL_THETA_200to400_ave.dat", skip = 7, header = FALSE)
#r16_CTRL <- read.delim("CTRL_THETA_200to980_ave.dat", skip = 7, header = FALSE)
#r17_CTRL <- read.delim("CTRL_THETA_250to500_ave.dat", skip = 7, header = FALSE)
#r18_CTRL <- read.delim("CTRL_THETA_400to1200_ave.dat", skip = 7, header = FALSE)

# Load experiment data, skipping the 7-line header
#r1 <- read.delim("GFDL_THETA_0to20_ave.dat", skip = 7, header = FALSE)
#r2 <- read.delim("GFDL_THETA_0to50_ave.dat", skip = 7, header = FALSE)
#r3 <- read.delim("GFDL_THETA_0to85_ave.dat", skip = 7, header = FALSE)
#r4 <- read.delim("GFDL_THETA_0to100_ave.dat", skip = 7, header = FALSE)
#r5 <- read.delim("GFDL_THETA_0to150_ave.dat", skip = 7, header = FALSE)
#r6 <- read.delim("GFDL_THETA_0to200_ave.dat", skip = 7, header = FALSE)
#r7 <- read.delim("GFDL_THETA_0to1200_ave.dat", skip = 7, header = FALSE)
#r8 <- read.delim("GFDL_THETA_50to150_ave.dat", skip = 7, header = FALSE)
#r9 <- read.delim("GFDL_THETA_50to200_ave.dat", skip = 7, header = FALSE)
#r10 <- read.delim("GFDL_THETA_50to300_ave.dat", skip = 7, header = FALSE)
#r11 <- read.delim("GFDL_THETA_100to200_ave.dat", skip = 7, header = FALSE)
#r12 <- read.delim("GFDL_THETA_100to250_ave.dat", skip = 7, header = FALSE)
#r13 <- read.delim("GFDL_THETA_100to400_ave.dat", skip = 7, header = FALSE)
#r14 <- read.delim("GFDL_THETA_150to600_ave.dat", skip = 7, header = FALSE)
#r15 <- read.delim("GFDL_THETA_200to400_ave.dat", skip = 7, header = FALSE)
#r16 <- read.delim("GFDL_THETA_200to980_ave.dat", skip = 7, header = FALSE)
#r17 <- read.delim("GFDL_THETA_250to500_ave.dat", skip = 7, header = FALSE)
#r18 <- read.delim("GFDL_THETA_400to1200_ave.dat", skip = 7, header = FALSE)

# Match with other therMizer parameters
#r <- c("r0to20", "r0to50", "r0to85", "r0to100", "r0to150", "r0to200",  "r0to1200", "r50to150", "r50to200", "r50to300", "r100to200", "r100to250", "r100to400", "r150to600", "r200to400", "r200to980", "r250to500", "r400to1200")

# Create monthly time steps, Jan 1961 - Dec 2010
#time_steps <- seq((1961 + 1/24), (2010 + 23/24), 1/12)

# Create ocean_temp array
#ocean_temp_realm_array <- array(NA, dim = c(length(time_steps), length(r)), dimnames = list(time = time_steps, realm_names = r))

# Fill array
#ocean_temp_realm_array[,] <- cbind((r1[,2] - mean(r1_CTRL[,2])) + r1_WOA[1,1],
                               (r2[,2] - mean(r2_CTRL[,2])) + r2_WOA[1,1],
                               (r3[,2] - mean(r3_CTRL[,2])) + r3_WOA[1,1],
                               (r4[,2] - mean(r4_CTRL[,2])) + r4_WOA[1,1],
                               (r5[,2] - mean(r5_CTRL[,2])) + r5_WOA[1,1],
                               (r6[,2] - mean(r6_CTRL[,2])) + r6_WOA[1,1],
                               (r7[,2] - mean(r7_CTRL[,2])) + r7_WOA[1,1],
                               (r8[,2] - mean(r8_CTRL[,2])) + r8_WOA[1,1],
                               (r9[,2] - mean(r9_CTRL[,2])) + r9_WOA[1,1],
                               (r10[,2] - mean(r10_CTRL[,2])) + r10_WOA[1,1],
                               (r11[,2] - mean(r11_CTRL[,2])) + r11_WOA[1,1],
                               (r12[,2] - mean(r12_CTRL[,2])) + r12_WOA[1,1],
                               (r13[,2] - mean(r13_CTRL[,2])) + r13_WOA[1,1],
                               (r14[,2] - mean(r14_CTRL[,2])) + r14_WOA[1,1],
                               (r15[,2] - mean(r15_CTRL[,2])) + r15_WOA[1,1],
                               (r16[,2] - mean(r16_CTRL[,2])) + r16_WOA[1,1],
                               (r17[,2] - mean(r17_CTRL[,2])) + r17_WOA[1,1],
                               (r18[,2] - mean(r18_CTRL[,2])) + r18_WOA[1,1])

# Save data
#write.table(ocean_temp_realm_array, file = "GFDL_ocean_temp_realm_array.dat", quote = FALSE, row.names = TRUE, col.names = TRUE)

```


```{r}
# Creating Ocean_temp_array - using the data created in the rmarkdown "Avg_temp_data_create.Rmd" and combining it for model runs .


# Load the data 

r1_ctrl <- readRDS("GFDL_CTRL_THETA_0to300_ave.rds")
r2_ctrl <- readRDS("GFDL_CTRL_THETA_0to500_ave.rds")
r3_ctrl <- readRDS("GFDL_CTRL_THETA_0to400_ave.rds")
r4_ctrl <- readRDS("GFDL_CTRL_THETA_100to700_ave.rds")
r5_ctrl <- readRDS("GFDL_CTRL_THETA_200to1100_ave.rds")
r6_ctrl <- readRDS("GFDL_CTRL_THETA_150to800_ave.rds")
r7_ctrl <- readRDS("GFDL_CTRL_THETA_400to1000_ave.rds")
r8_ctrl <- readRDS("GFDL_CTRL_THETA_400to800_ave.rds")
r9_ctrl <- readRDS("GFDL_CTRL_THETA_400to600_ave.rds")
r10_ctrl <- readRDS("GFDL_CTRL_THETA_100to800_ave.rds")
r11_ctrl <- readRDS("GFDL_CTRL_THETA_100to400_ave.rds")
r12_ctrl <- readRDS("GFDL_CTRL_THETA_50to300_ave.rds")
r13_ctrl <- readRDS("GFDL_CTRL_THETA_150to400_ave.rds")
r14_ctrl <- readRDS("GFDL_CTRL_THETA_50to500_ave.rds")
r15_ctrl <- readRDS("GFDL_CTRL_THETA_50to400_ave.rds")



# Create a coulmn of all the realm names 
realms <- c("r0-300","r0-500","r0-400","r100-700","r200-1100","r150-800","r400-1000","r400-800","r400-600","r100-800","r100-400","r50-300","r150-400","r50-500","r50-400")

#Create yearly time steps, 1961 - 2010
years <- 1961:2010
time_steps <- years

Ocean_temp_realm_array <- array(NA, dim = c(length(time_steps), length(realms)), dimnames = list(time = time_steps, realm_names = realms))

# Fill the array by combing all the above data 

Ocean_temp_realm_array[,] <- cbind(r1_ctrl[,2], r2_ctrl[,2], r3_ctrl[,2], r4_ctrl[,2], r5_ctrl[,2], r6_ctrl[,2], r7_ctrl[,2], r8_ctrl[,2], r9_ctrl[,2], r10_ctrl[,2], r11_ctrl[,2], r12_ctrl[,2], r13_ctrl[,2], r14_ctrl[,2], r15_ctrl[,2])


#write.table(Ocean_temp_realm_array, file = "SEAus_Ocean_temp_realm_array.dat", quote = FALSE, row.names = TRUE, col.names = TRUE)

#saveRDS(Ocean_temp_realm_array, "Ocean_temp_realm_array.rds")

```
