==> 04_Model_runs.Rmd <==
---
title: "Model_Runs"
output: html_document
date: "2024-04-02"
editor_options: 
  chunk_output_type: console
---

## Model runs

This script runs the two simulations: one with (histsoc) and one without (nat) fishing.

Set up the environment:
```{r}
# Load libraries
library("remotes")
remotes::install_github("sizespectrum/therMizer")
library("therMizer")
```

Set up the species parameters:
```{r}
# Load parameters 
Sea_params <- readRDS("final_params_object.rds")
Sea_params@species_params$temp_min <- c(13,14,12,2,2,14,4,11,12,9,11,3,8,9,8,3,14,14,6)
Sea_params@species_params$temp_max <-c(29,20,23,15,16,21,19,16,19,29,20,9,16,16,14,13,18,20,24)

interaction <- Sea_params@interaction
interaction <- as.matrix(interaction)
params <- newMultispeciesParams(species_params = Sea_params@species_params,interaction = interaction, kappa = 0.002191539)
```

Set up time steps:
```{r}
#time_steps <- seq(0,600,1)
time_steps<- 1961:2010
# I ran into trouble using any time steps that didn't start with 0.
# I also had trouble using fractional time steps.  
# So, what's here as year is really a month.  
# Suggestions for a better solution welcome
```

Set up the resource:
```{r}
# Load resource spectra 
# Note: You may need to update this path
isimip_plankton <- read.table("SEAus_GFDL_resource_spectra_S1.03I0.85.dat")
isimip_plankton <- as(isimip_plankton, "matrix")
sizes <- names(params@initial_n_pp)
n_pp_array <- array(NA, dim = c(length(time_steps), length(sizes)), dimnames = list(time = time_steps, w = sizes))

# Fill array
# Need an extra time step preceding the simulation
n_pp_array[1,] <- isimip_plankton[1,]
for (t in seq(1,length(time_steps) - 1,1)) {
  n_pp_array[t + 1,] <- isimip_plankton[t,]
}
```

Set up temperature-related information:
```{r}
# Load ocean temp array

ocean_temp <- readRDS("Ocean_temp_realm_array.rds")
ocean_temp <- as(ocean_temp, "matrix")
r <- colnames(ocean_temp)
ocean_temp_array <- array(NA, dim = c(length(time_steps), length(r)), dimnames = list(time = time_steps, realm_names = r))

ocean_temp_array[1,] <- ocean_temp[1,]
for (t in seq(1,length(time_steps) - 1, 1)) {
  ocean_temp_array[t + 1,] <- ocean_temp[t,]
}

# Load vertical migration and exposure
load("SEAus_vertical_migration_array.Rdata") # Note: this is named vertical_migration_array

Sea_exposure <- read.table("SEAus_exposure_array.dat")
Sea_exposure <- as(Sea_exposure, "matrix")
```

Set up the needed fishing selectivity function:
```{r}
# We need to define our double knife-edge selectivity function to match that 
# used in Woodworth-Jefcoats et al. 2019
#double_knife_edge <- function(w, knife_edge_size1, knife_edge_size2, ...) { 
	#sel <- rep(0, length(w)) 
	
	# Phased in linearly
	#F0 <- which(w < knife_edge_size1) # to find one size smaller than that fished, for the 0 value 
	#F1 <- which(w < knife_edge_size2) # to find end of escalation size range 
	#lo_sel <- max(F0):max(F1) 
	#sel[lo_sel] <- seq(0, 1, length = length(lo_sel)) # linear increase from 0 to F 
	#sel[w >= knife_edge_size2] <- 1 
	#return(sel) 
#} 
```

Add gear parameters:
```{r}
#gear_params(params) <- HIgears
```

Load effort:
```{r}
# Load effort
# Note: You may need to update this path
isimip_effort <- readRDS("Fishing_his.rds")
isimip_effort <- as(isimip_effort, "matrix")
gear_name = c("pelagic", "demersal")
effort_array <- array(NA, c(length(time_steps), length(gear_name)), dimnames = list(time = time_steps, gear = gear_name))
effort_array[1,] = isimip_effort[1,2]
for (t in seq(1,length(time_steps) - 1, 1)) {
  effort_array[t + 1] <- isimip_effort[t,2]
}
```

Add temperature and resource information to the parameters:
```{r}
params <- upgradeTherParams(params, ocean_temp_array = ocean_temp_array,
                            n_pp_array = n_pp_array, 
                            vertical_migration_array = vertical_migration_array,
                            exposure_array = Sea_exposure,
                            aerobic_effect = TRUE, metabolism_effect = TRUE)

# Nice test to see that things are working correctly
plotThermPerformance(params)
plot(params)
```

Project to steady-state, with and without fishing:
```{r}
# Project to steady
params_steady_F <- projectToSteady(params, t_max = 600, dt = 1/12, tol = 1e-7, effort = effort_array[1]) 
sim_steady_F <- projectToSteady(params, t_max = 600, dt = 1/12, tol = 1e-7, effort = effort_array[1], return_sim = TRUE) 
params_steady_noF <- projectToSteady(params, t_max = 600, dt = 1/12, tol = 1e-7, effort = 0) 
sim_steady_noF <- projectToSteady(params, t_max = 600, dt = 1/12, tol = 1e-7, effort = 0, return_sim = TRUE) 

# Include a validation run with F = 0.2 (the estimated F for most of the species modeled, per stock assessments)
params_steady_validate <- projectToSteady(params, t_max = 600, dt = 1/12, tol = 1e-7, effort = 0.2) 
sim_steady_validate <- projectToSteady(params, t_max = 600, dt = 1/12, tol = 1e-7, effort = 0.2, return_sim = TRUE) 

# Test to see how things look
plot(sim_steady_F)
plot(sim_steady_noF)
plot(sim_steady_validate)
```

Run simulations:
```{r}
# And do a simulation using the steady parameters
projection_steady_F <- project(params_steady_F, dt = 1/12, t_max = 600, t_save = 1/12, effort = effort_array)
projection_steady_noF <- project(params_steady_noF, dt = 1/12, t_max = 600, t_save = 1/12, effort = 0)

# Test to see how things look
plot(projection_steady_F)
plot(projection_steady_noF)
```

Save output:
```{r}
# Save projections for investigating results and formatting for ISIMIP
save(sim_steady_validate, file = "sim_steady_validate.Rdata")
save(projection_steady_F, file = "projection_steady_F.Rdata")
save(projection_steady_noF, file = "projection_steady_noF.Rdata")
```
